adds responsive download packet
